Initial commit
This puts in communication with the command interface of the laser
controllers, providing that no errors ever occur.

Two interfaces are available, the higher-level Python interface provided
in the root of the package namespace (defined in `instrument.py`), and
the lower-level wrapper around the telnet connection.

Stubs are in place for error handling and the future ReactiveX
`Observable` interface to the monitoring system, but these aren't yet
written.